Separated kernel and user code
