+++ src/kernel/interrupts.rs
@@ -1,4 +1,6 @@
-use crate::{clock, gdt, print, shell};
+use crate::kernel::{clock, gdt};
+use crate::print;
+use crate::user;
 use lazy_static::lazy_static;
 use pic8259::ChainedPics;
 use spin;
@@ -66,7 +68,7 @@ extern "x86-interrupt" fn timer_interrupt_handler(_stack_frame: InterruptStackFr
 }
 
 extern "x86-interrupt" fn keyboard_interrupt_handler(_stack_frame: InterruptStackFrame) {
-    use crate::layouts;
+    use crate::kernel::layouts;
     use pc_keyboard::{DecodedKey, HandleControl, Keyboard, ScancodeSet1};
     use spin::Mutex;
     use x86_64::instructions::port::Port;
@@ -86,8 +88,9 @@ extern "x86-interrupt" fn keyboard_interrupt_handler(_stack_frame: InterruptStac
     let scancode: u8 = unsafe { port.read() };
     if let Ok(Some(key_event)) = keyboard.add_byte(scancode) {
         if let Some(key) = keyboard.process_keyevent(key_event) {
+            //user::shell::key_handle(key);
             match key {
-                DecodedKey::Unicode(c) => shell::key_handle(c),
+                DecodedKey::Unicode(c) => user::shell::key_handle(c),
                 DecodedKey::RawKey(_) => {}
             }
         }
