+++ src/kernel/mod.rs
@@ -1,0 +1,6 @@
+pub mod clock;
+pub mod gdt;
+pub mod interrupts;
+pub mod layouts;
+pub mod string;
+pub mod vga;
