+++ src/kernel/string.rs
